Remove final references to captai
