+++ ironfish/src/testUtilities/fake/helpers/index.ts
@@ -2,9 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-export * from './anchorChain'
 export * from './blockchain'
-export * from './captain'
 export * from './jest'
 export * from './merkleTree'
 export * from './storage'
